qa: Move cli.rs to config/
